1. Add NaN warning to EnergyTracker2. 2. Add docs on flow analysis 3. Add GlUtils::AlignedBOxWithTicks 4. Add analysis functions to FlowAnalysis 5. Fix elastic energy computation for adhesive hertzian models (was NaN due to pow(negative_overlap,2/5.)) 6. Move several classes in pkg/dem/Particle.hpp to new-style macros 7. Add converter for vector<size_t> from python 8. Fix HTML dump of AlignedBox3 (bug in minigien, return len as 3, although it has only 2, min and max)

+++ core/Test.cpp
@@ -57,6 +57,6 @@ py::object WooTestClass::arr3d_py_get(){
 void WooTestClass::arr3d_set(const Vector3i& shape, const VectorXr& data){
 	if(data.rows()!=shape.prod()) throw std::runtime_error("Data must have exactly "+to_string(shape[0])+"x"+to_string(shape[1])+"x"+to_string(shape[2])+"="+to_string(shape.prod())+" items (not "+to_string(data.rows())+").");
 	arr3d.resize(boost::extents[shape[0]][shape[1]][shape[2]]);
-	for(size_t i=0; i<shape[0]; i++) for(size_t j=0; j<shape[1]; j++) for(size_t k=0; k<shape[2]; k++) arr3d[i][j][k]=data[i*shape[0]*shape[1]+j*shape[1]+k];
+	for(int i=0; i<shape[0]; i++) for(int j=0; j<shape[1]; j++) for(int k=0; k<shape[2]; k++) arr3d[i][j][k]=data[i*shape[0]*shape[1]+j*shape[1]+k];
 }
